refactor: Rename classes/ to classlevel/
