+++ src/winter/util/AnnotationScanner.java
@@ -9,7 +9,7 @@ import java.util.Enumeration;
 import java.util.Map;
 
 import jakarta.servlet.ServletContext;
-import winter.annotation.classes.Controller;
+import winter.annotation.classlevel.Controller;
 import winter.annotation.methods.UrlMapping;
 import winter.data.Mapping;
 import winter.data.MappingMethod;
